+++ mmxix/src/be/strijbol/advent/mmxix/day7/Main.java
@@ -18,16 +18,16 @@ import java.util.stream.StreamSupport;
  */
 public class Main {
 
-    private static int one() throws FileNotFoundException {
+    private static long one() throws FileNotFoundException {
         var program = Inputs.line(2019, 7);
         var computer = new Computer();
         // There are 5 thrusters, so try every combination without repetition for 0-4.
-        var possibleValues = List.of(0, 1, 2, 3, 4);
+        var possibleValues = List.of(0L, 1L, 2L, 3L, 4L);
         return StreamSupport.stream(PermutationIterator.spliterator(possibleValues), false)
-                .mapToInt(inputs -> {
-                    var input = 0;
+                .mapToLong(inputs -> {
+                    var input = 0L;
                     for (var phase : inputs) {
-                        var out = new CollectingConsumer<Integer>(false);
+                        var out = new CollectingConsumer<Long>(false);
                         computer.loadProgram(program);
                         computer.execute(Lists.asSupplier(List.of(phase, input)), out);
                         input = out.getElement().orElseThrow();
@@ -37,7 +37,7 @@ public class Main {
                 .max().orElse(0);
     }
 
-    private static int executeFeedbackLoop(List<Integer> phases) {
+    private static long executeFeedbackLoop(List<Long> phases) {
         var amplifiers = new ArrayList<Amplifier>();
         String program;
         try {
@@ -45,7 +45,7 @@ public class Main {
         } catch (FileNotFoundException e) {
             throw new RuntimeException(e); // Prevent errors.
         }
-        Consumer<Integer> noOutput = integer -> { throw new HaltException(); };
+        Consumer<Long> noOutput = integer -> { throw new HaltException(); };
         // Set the phases.
         Amplifier previous = null;
         for (var phase : phases) {
@@ -61,22 +61,22 @@ public class Main {
             amplifiers.add(amplifier);
         }
 
-        var finalResult = new CollectingConsumer<Integer>(true);
+        var finalResult = new CollectingConsumer<Long>(true);
         var last = amplifiers.get(amplifiers.size() - 1);
         var first = amplifiers.get(0);
-        Consumer<Integer> thrusterAndCallback = integer -> {
+        Consumer<Long> thrusterAndCallback = integer -> {
             finalResult.accept(integer);
             first.accept(integer);
         };
         last.setOutput(thrusterAndCallback);
-        first.accept(0);
+        first.accept(0L);
         return finalResult.getElement().orElseThrow();
     }
 
-    private static int two() {
-        var possibleValues = List.of(5, 6, 7, 8, 9);
+    private static long two() {
+        var possibleValues = List.of(5L, 6L, 7L, 8L, 9L);
         return StreamSupport.stream(PermutationIterator.spliterator(possibleValues), false)
-                .mapToInt(Main::executeFeedbackLoop)
+                .mapToLong(Main::executeFeedbackLoop)
                 .max()
                 .orElseThrow();
     }
